sre-copilot/pyproject.toml
[project]
name = "sre-copilot"
version = "0.1.0"
description = "AI SRE Copilot scaffolding with gateway, guardrails, routing, and retrieval layers."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.9.1",
    "nemoguardrails>=0.10.0",
    "httpx>=0.27.2",
    "openai>=1.55.3",
    "pymilvus>=2.4.6,<2.5.0",
    "redis>=5.0.0,<6.0.0",
    "opentelemetry-sdk>=1.28.2",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.2",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-requests>=0.48b0",
    "requests>=2.32.3",
    "langfuse>=2.46.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.7.4",
    "respx>=0.20.2",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"sre_copilot" = ["guardrails/config/config.yaml"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
